// src/types.ts
export interface CatCan {
  made: string // 產地
  brand: string
  name: string
  flaver: string
  weight_g: string
  kcal: string
  protein: string
  fat: string
  fiber: string
  ash: string
  moisture: string // 水分百分比 (%)
  moistureContent: string // 水分含量 (ml)
  Ca: string
  P: string
  taurine: string // 牛磺酸濃度 (mg/kg)
  taurineContent: string // 牛磺酸含量 (mg)
}